refactor: move ensure_midenup_home_exists logic to init
